server/requirements.txt
openenv-core>=0.2.0
uvicorn>=0.30.0
fastapi>=0.115.0
pydantic>=2.0.0
